fix libjvm copy command on JDK8

+++ pkgbuild/pkgbuild.sh
@@ -67,8 +67,11 @@ done
 
 rm -rf *.pkg distribution.xml Resources/en.lproj/welcome.html Resources/en.lproj/conclusion.html
 mkdir -p "${INPUT_DIRECTORY}/Contents/Home/bundle/Libraries"
-cp "${INPUT_DIRECTORY}/Contents/Home/lib/server/libjvm.dylib" "${INPUT_DIRECTORY}/Contents/Home/bundle/Libraries/libserver.dylib"
-
+if [ -f "${INPUT_DIRECTORY}/Contents/Home/lib/server/libjvm.dylib" ]; then
+  cp "${INPUT_DIRECTORY}/Contents/Home/lib/server/libjvm.dylib" "${INPUT_DIRECTORY}/Contents/Home/bundle/Libraries/libserver.dylib"
+else
+  cp "${INPUT_DIRECTORY}/Contents/Home/jre/lib/server/libjvm.dylib" "${INPUT_DIRECTORY}/Contents/Home/bundle/Libraries/libserver.dylib"
+fi
 # Detect if JRE or JDK
 case $INPUT_DIRECTORY in
   *-jre)
